+++ users.go
@@ -6,28 +6,56 @@ import (
 	"net/http"
 )
 
+// CreateUserRequest is the set of data required to create a new user in MagicBell.
+// Please note that you must provide the user's email or the external id so MagicBell can uniquely identify the user.
+// The external id, if provided, must be unique to the user.
 type CreateUserRequest struct {
-	ExternalID       string            `json:"external_id"`
-	Email            string            `json:"email"`
-	FirstName        string            `json:"first_name"`
-	LastName         string            `json:"last_name"`
-	CustomAttributes map[string]string `json:"custom_attributes"`
+	// ExternalID is a unique string that MagicBell can utilize to uniquely identify the user.
+	// We recommend setting this attribute to the ID of the user in your database.
+	// Provide the external id if the user's email is unavailable.
+	ExternalID string `json:"external_id"`
+	// Email is the user's email.
+	Email string `json:"email"`
+	// FirstName is the user's first name.
+	FirstName string `json:"first_name,omitempty"`
+	// LastName is the user's last name.
+	LastName string `json:"last_name,omitempty"`
+	// CustomAttributes are any customers attributes that you'd like to associate with the user.
+	// These custom attributes can later be utilized in MagicBell's web interface (when writing email templates for example).
+	CustomAttributes map[string]string `json:"custom_attributes,omitempty"`
 }
 
+// UpdateUserRequest is the set of data required to update an existing user in MagicBell.
+// Note that this is a PUT request, so all data given will overwrite all existing data in MagicBell.
 type UpdateUserRequest struct {
-	ExternalID       string            `json:"external_id"`
-	Email            string            `json:"email"`
-	FirstName        string            `json:"first_name"`
-	LastName         string            `json:"last_name"`
-	CustomAttributes map[string]string `json:"custom_attributes"`
+	// ExternalID is a unique string that MagicBell can utilize to uniquely identify the user.
+	// We recommend setting this attribute to the ID of the user in your database.
+	// Provide the external id if the user's email is unavailable.
+	ExternalID string `json:"external_id"`
+	// Email is the user's email.
+	Email string `json:"email"`
+	// FirstName is the user's first name.
+	FirstName string `json:"first_name,omitempty"`
+	// LastName is the user's last name.
+	LastName string `json:"last_name,omitempty"`
+	// CustomAttributes are any customers attributes that you'd like to associate with the user.
+	// These custom attributes can later be utilized in MagicBell's web interface (when writing email templates for example).
+	CustomAttributes map[string]string `json:"custom_attributes,omitempty"`
 }
 
+// User is MagicBell's representation of a user, uniquely identified by ID.
 type User struct {
-	ID               string            `json:"id"`
-	ExternalID       string            `json:"external_id"`
-	Email            string            `json:"email"`
-	FirstName        string            `json:"first_name"`
-	LastName         string            `json:"last_name"`
+	// ID is the unique ID MagicBell internally uses to represent your user.
+	ID string `json:"id"`
+	// ExternalID is a unique string that MagicBell can utilize to uniquely identify the user.
+	ExternalID string `json:"external_id"`
+	// Email is the user's email.
+	Email string `json:"email"`
+	// FirstName is the user's first name.
+	FirstName string `json:"first_name"`
+	// LastName is the user's last name.
+	LastName string `json:"last_name"`
+	// CustomAttributes are any customers attributes that you'd like to associate with the user.
 	CustomAttributes map[string]string `json:"custom_attributes"`
 }